+++ crates/holochain_integrity_types/src/action.rs
@@ -1,20 +1,10 @@
 use crate::entry_def::EntryVisibility;
-use crate::link::LinkTag;
-use crate::link::LinkType;
-use crate::EntryRateWeight;
-use crate::MembraneProof;
-use crate::RateWeight;
 use holo_hash::ActionHash;
 use holo_hash::AgentPubKey;
-use holo_hash::AnyLinkableHash;
 use holo_hash::DnaHash;
-use holo_hash::EntryHash;
 use holochain_serialized_bytes::prelude::*;
-use holochain_timestamp::Timestamp;
 use std::borrow::Borrow;
-use std::hash::Hash;
 
-pub mod builder;
 pub mod conversions;
 
 /// Any action with a action_seq less than this value is part of a record
@@ -47,59 +37,16 @@ macro_rules! write_into_action {
     };
 }
 
-/// A trait to unify the "inner" parts of an Action, i.e. the structs inside
-/// the Action enum's variants. This trait is used for the "weighed" version
-/// of each struct, i.e. the version without weight information erased.
-///
-/// Action types with no weight are considered "weighed" and "unweighed" at the
-/// same time, but types with weight have distinct types for the weighed and
-/// unweighed versions.
-pub trait ActionWeighed {
-    type Unweighed: ActionUnweighed;
-    type Weight: Default;
-
-    /// Erase the rate limiting weight info, creating an "unweighed" version
-    /// of this action. This is used primarily by validators who need to run the
-    /// `weigh` callback on an action they received, and want to make sure their
-    /// callback is not using the predefined weight to influence the result.
-    fn unweighed(self) -> Self::Unweighed;
-}
-
-/// A trait to unify the "inner" parts of an Action, i.e. the structs inside
-/// the Action enum's variants. This trait is used for the "unweighed" version
-/// of each struct, i.e. the version with weight information erased.
-///
-/// Action types with no weight are considered "weighed" and "unweighed" at the
-/// same time, but types with weight have distinct types for the weighed and
-/// unweighed versions.
-pub trait ActionUnweighed: Sized {
-    type Weighed: ActionWeighed;
-    type Weight: Default;
-
-    /// Add a weight to this unweighed action, making it "weighed".
-    /// The weight is determined by the `weigh` callback, which is run on the
-    /// unweighed version of this action.
-    fn weighed(self, weight: Self::Weight) -> Self::Weighed;
-
-    /// Add zero weight to this unweighed action, making it "weighed".
-    #[cfg(feature = "test_utils")]
-    fn weightless(self) -> Self::Weighed {
-        self.weighed(Default::default())
-    }
-}
-
 write_into_action! {
     Dna,
     AgentValidationPkg,
     InitZomesComplete,
     OpenChain,
     CloseChain,
-
-    Create<EntryRateWeight>,
-    Update<EntryRateWeight>,
-    Delete<RateWeight>,
-
-    CreateLink<RateWeight>,
+    Create,
+    Update,
+    Delete,
+    CreateLink,
     DeleteLink,
 }
 
@@ -142,71 +89,6 @@ impl ZomeIndex {
 )]
 pub struct EntryDefIndex(pub u8);
 
-/// The Dna Action is always the first action in a source chain
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, SerializedBytes, Hash)]
-pub struct Dna {
-    pub author: AgentPubKey,
-    pub timestamp: Timestamp,
-    // No previous action, because DNA is always first chain entry
-    pub hash: DnaHash,
-}
-
-/// Action for an agent validation package, used to determine whether an agent
-/// is allowed to participate in this DNA
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, SerializedBytes, Hash)]
-pub struct AgentValidationPkg {
-    pub author: AgentPubKey,
-    pub timestamp: Timestamp,
-    pub action_seq: u32,
-    pub prev_action: ActionHash,
-
-    pub membrane_proof: Option<MembraneProof>,
-}
-
-/// An action which declares that all zome init functions have successfully
-/// completed, and the chain is ready for commits. Contains no explicit data.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, SerializedBytes, Hash)]
-pub struct InitZomesComplete {
-    pub author: AgentPubKey,
-    pub timestamp: Timestamp,
-    pub action_seq: u32,
-    pub prev_action: ActionHash,
-}
-
-/// Declares that a metadata Link should be made between two hashes of anything; could be data or
-/// an op or anything that can be hashed.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, SerializedBytes, Hash)]
-pub struct CreateLink<W = RateWeight> {
-    pub author: AgentPubKey,
-    pub timestamp: Timestamp,
-    pub action_seq: u32,
-    pub prev_action: ActionHash,
-
-    pub base_address: AnyLinkableHash,
-    pub target_address: AnyLinkableHash,
-    pub zome_index: ZomeIndex,
-    pub link_type: LinkType,
-    pub tag: LinkTag,
-
-    pub weight: W,
-}
-
-/// Declares that a previously made Link should be nullified and considered removed.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, SerializedBytes, Hash)]
-pub struct DeleteLink {
-    pub author: AgentPubKey,
-    pub timestamp: Timestamp,
-    pub action_seq: u32,
-    pub prev_action: ActionHash,
-
-    /// this is redundant with the `CreateLink` action but needs to be included to facilitate DHT ops
-    /// this is NOT exposed to wasm developers and is validated by the subconscious to ensure that
-    /// it always matches the `base_address` of the `CreateLink`
-    pub base_address: AnyLinkableHash,
-    /// The address of the `CreateLink` being reversed
-    pub link_add_address: ActionHash,
-}
-
 /// Description of how to find the previous or next CellId in a migration.
 /// In a migration, of the two components of the CellId (dna and agent),
 /// always one stays fixed while the other one changes.
@@ -232,103 +114,6 @@ impl From<AgentPubKey> for MigrationTarget {
     fn from(agent: AgentPubKey) -> Self {
         MigrationTarget::Agent(agent)
     }
-}
-
-/// When migrating to a new version of a DNA, this action is committed to the
-/// old chain to declare the migration path taken. This action can also be taken
-/// to simply close down a chain with no forward reference to a migration.
-///
-/// Note that if `MigrationTarget::Agent` is used, this action will be signed with
-/// that key rather than the authoring key, so that new key must be a valid keypair
-/// that you control in the keystore, so that the action can be signed.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, SerializedBytes, Hash)]
-pub struct CloseChain {
-    pub author: AgentPubKey,
-    pub timestamp: Timestamp,
-    pub action_seq: u32,
-    pub prev_action: ActionHash,
-
-    pub new_target: Option<MigrationTarget>,
-}
-
-/// When migrating to a new version of a DNA, this action is committed to the
-/// new chain to declare the migration path taken.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, SerializedBytes, Hash)]
-pub struct OpenChain {
-    pub author: AgentPubKey,
-    pub timestamp: Timestamp,
-    pub action_seq: u32,
-    pub prev_action: ActionHash,
-
-    pub prev_target: MigrationTarget,
-    /// The hash of the `CloseChain` action on the old chain, to establish chain continuity
-    /// and disallow backlinks to multiple forks on the old chain.
-    pub close_hash: ActionHash,
-}
-
-/// An action which "speaks" Entry content into being. The same content can be
-/// referenced by multiple such actions.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, SerializedBytes, Hash)]
-pub struct Create<W = EntryRateWeight> {
-    pub author: AgentPubKey,
-    pub timestamp: Timestamp,
-    pub action_seq: u32,
-    pub prev_action: ActionHash,
-
-    pub entry_type: EntryType,
-    pub entry_hash: EntryHash,
-
-    pub weight: W,
-}
-
-/// An action which specifies that some new Entry content is intended to be an
-/// update to some old Entry.
-///
-/// This action semantically updates an entry to a new entry.
-/// It has the following effects:
-/// - Create a new Entry
-/// - This is the action of that new entry
-/// - Create a metadata relationship between the original entry and this new action
-///
-/// The original action is required to prevent update loops:
-/// If you update entry A to B and B back to A, and only track the original entry,
-/// then you have a loop of references. Every update introduces a new action,
-/// so there can only be a linear history of action updates, even if the entry history
-/// experiences repeats.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, SerializedBytes, Hash)]
-pub struct Update<W = EntryRateWeight> {
-    pub author: AgentPubKey,
-    pub timestamp: Timestamp,
-    pub action_seq: u32,
-    pub prev_action: ActionHash,
-
-    pub original_action_address: ActionHash,
-    pub original_entry_address: EntryHash,
-
-    pub entry_type: EntryType,
-    pub entry_hash: EntryHash,
-
-    pub weight: W,
-}
-
-/// Declare that a previously published Action should be nullified and
-/// considered deleted.
-///
-/// Via the associated [`crate::dht_v2::op::Op`], this also has an effect on Entries: namely,
-/// that a previously published Entry will become inaccessible if all of its
-/// Actions are marked deleted.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, SerializedBytes, Hash)]
-pub struct Delete<W = RateWeight> {
-    pub author: AgentPubKey,
-    pub timestamp: Timestamp,
-    pub action_seq: u32,
-    pub prev_action: ActionHash,
-
-    /// Address of the Record being deleted
-    pub deletes_address: ActionHash,
-    pub deletes_entry_address: EntryHash,
-
-    pub weight: W,
 }
 
 /// Allows Actions which reference Entries to know what type of Entry it is