+++ crates/holochain_types/src/prelude.rs
@@ -1,7 +1,6 @@
 //! reexport some common things
 
 pub use crate::access::*;
-pub use crate::action::*;
 pub use crate::activity::*;
 pub use crate::app::*;
 pub use crate::chain::*;